requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf~=3.19.0
chromadb
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
